x2

+++ Android/App/app/src/main/java/rocky/rockyapp/app/Tools.java
@@ -30,12 +30,12 @@ class Tools {
         return (int)((gyroValue + 0.5) * 1024 );
     }
 
-    static public int gyroToServo256(float gyroValue){
+    static public byte gyroToServo256(float gyroValue){
         if(gyroValue >= 0.5)
-            return 255;
+            return (byte)255;
         if(gyroValue <= -0.5)
             return 0;
-        return (int)((gyroValue + 0.5) * 256 );
+        return (byte)(int)((gyroValue + 0.5) * 256 );
     }
 
     static public int gyroToServoInverse(float gyroValue){ //Inversa, por si el servo gira al reves, solo dejar la que se valla a usar
